Moved actions to services folder. Email to lowercased in db Clean up
