+++ nextjs/src/app/auth/login/page.tsx
@@ -1,4 +1,4 @@
-import LoginForm from '@/components/LoginForm';
+import LoginForm from '@components/Auth/LoginForm';
 
 export default function Login() {
   return <LoginForm />;
