+++ nextjs/src/app/auth/new-password/page.tsx
@@ -1,4 +1,4 @@
-import NewPasswordForm from '@/components/NewPasswordForm';
+import NewPasswordForm from '@components/Auth/NewPasswordForm';
 
 export default function NewVerification() {
   return <NewPasswordForm/>;
