+++ nextjs/src/app/auth/new-verification/page.tsx
@@ -1,4 +1,4 @@
-import VerificationForm from '@/components/VerificationForm';
+import VerificationForm from '@components/Auth/VerificationForm';
 
 export default function NewVerification() {
   return <VerificationForm />;
