+++ nextjs/src/app/auth/register/page.tsx
@@ -1,4 +1,4 @@
-import RegisterForm from '@/components/RegisterForm';
+import RegisterForm from '@components/Auth/RegisterForm';
 import React from 'react';
 
 export default function Register() {
