--- nextjs/src/model/Types.ts
@@ -1,3 +1,0 @@
-export type TokenType = 'VerifyEmail' | 'ForgotPassword'
-
-export type Gender = 'Male' | 'Female' | 'Other'
