+++ nextjs/src/services/Auth/EmailService/index.ts
